requirements.txt
asgiref==3.8.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-filter==25.1
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.34.1
MarkupSafe==3.0.2
msgpack==1.1.0
oauthlib==3.2.2
psutil==7.0.0
psycopg==3.2.6
pycparser==2.22
PyJWT==2.9.0
pyzmq==26.4.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==78.1.0
sqlparse==0.5.3
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
